big refactor, bug fix, and feature additions to pipeline.py.
the incoming dataset is now setified (setify=False) to ensure all claims are kept.
much better logging, showing metrics for missed responses/hits, and providing example outputs for each step (queries, predictions, etc ...)
spacy nlp on claims and articles is now done beforehand in batch via nlp.pipe using the n_process arg for multiprocessing.
multiprocessing for fetching responses and generating support was removed due to headaches :(, damn you multiprocessing.
pipeline was refactored to return resources in dicts where the keys are the claim id or article id, and the values were the computed resource. instead of utilizing these dicts, we update the claims/articles with the resource direclty by adding it as an attribute. this is a lot cleaner and can be done in the main part of the script for a clearer understanding of how the data is being updated at each step.